+++ inc/document_item.class.php
@@ -574,8 +574,10 @@ class Document_Item extends CommonDBRelation{
 
       // find documents already associated to the item
       $doc_item   = new self();
-      $used_found = $doc_item->find("`items_id` = '".$item->getID()."'
-                                    AND `itemtype` = '".$item->getType()."'");
+      $used_found = $doc_item->find([
+         'items_id'  => $item->getID(),
+         'itemtype'  => $item->getType()
+      ]);
       $used       = array_keys($used_found);
       $used       = array_combine($used, $used);