+++ inc/knowbaseitemtranslation.class.php
@@ -205,7 +205,7 @@ class KnowbaseItemTranslation extends CommonDBChild {
       }
 
       $obj   = new self;
-      $found = $obj->find("`knowbaseitems_id`='".$item->getID()."'", "`language` ASC");
+      $found = $obj->find(['knowbaseitems_id' => $item->getID()], "`language` ASC");
 
       if (count($found) > 0) {
          if ($canedit) {
@@ -330,8 +330,10 @@ class KnowbaseItemTranslation extends CommonDBChild {
       global $DB;
 
       $obj   = new self;
-      $found = $obj->find("`knowbaseitems_id` = '".$item->getID().
-                          "' AND `language` = '".$_SESSION['glpilanguage']."'");
+      $found = $obj->find([
+         'knowbaseitems_id'   => $item->getID(),
+         'language'           => $_SESSION['glpilanguage']
+      ]);
 
       if ((count($found) > 0)
           && in_array($field, ['name', 'answer'])) {