+++ inc/toolbox.class.php
@@ -2653,7 +2653,7 @@ class Toolbox {
          preg_match_all('/'.Document::getImageTag('(([a-z0-9]+|[\.\-]?)+)').'/', $content_text,
                         $matches, PREG_PATTERN_ORDER);
          if (isset($matches[1]) && count($matches[1])) {
-            $doc_data = $document->find("`tag` IN('".implode("','", array_unique($matches[1]))."')");
+            $doc_data = $document->find(['tag' => array_unique($matches[1])]);
          }
       }